Fix duplicate classes of spells
This was due to a missing deep copy of a mutable list.

This bug meant some spells with multiple classes, e.g.
stoneskin with druid, mage, monk, would be repeated
e.g. 3 times as a "monk" spell only. It wrongly chose
the final class in the classes list for the spell.